Company edit & create

+++ app/Models/Company.php
@@ -29,8 +29,8 @@ class Company extends Model
 
     protected $casts = [
         'open_hours' => 'array',
-        'latitude' => 'decimal',
-        'longitude' => 'decimal',
+        'latitude' => 'float',
+        'longitude' => 'float',
     ];
 
     public function owner(): BelongsTo
